requirements.txt
requests_oauthlib>=0.6.2
requests_toolbelt>=0.7.0
msgpack-python>=0.4.8